created MVC API for trailer details, truck details, its efficiency, control, and info
